+++ js/photos.js
@@ -2,7 +2,6 @@ const photoTemplate = document.querySelector('#picture')
   .content
   .querySelector('.picture');
 
-export const container = document.querySelector('.pictures');
 
 export const createFoto = ({ comments, description, likes, url, id}) => {
   const photo = photoTemplate.cloneNode(true);
@@ -14,16 +13,4 @@ export const createFoto = ({ comments, description, likes, url, id}) => {
   photo.dataset.photoId = id;
 
   return photo;
-};
-
-export const renderingPhoto = (pictures) => {
-  const fragment = document.createDocumentFragment();
-
-  pictures.forEach((picture) => {
-    const photo = createFoto(picture);
-
-    fragment.append(photo);
-  });
-
-  container.append(fragment);
 };
